pyproject.toml
[project]
name = "news-kg"
version = "0.1.0"
description = "News knowledge graph pipeline"
readme = "README.md"
requires-python = "==3.12.*"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "dspy-ai>=3.1.3",
    "gliner2>=1.2.4",
    "httpx>=0.28.1",
    "lxml>=6.0.2",
    "prefect>=3.6.20",
    "pydantic>=2.12.5",
    "typer>=0.24.1",
]

[project.scripts]
news-kg = "news_kg.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
markers = [
    "live: hits real external URLs (skip with -m 'not live')",
]

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
ignore = ["E501"]

[dependency-groups]
dev = [
    "ipykernel>=7.2.0",
    "jupyterlab>=4.5.5",
    "pre-commit>=4.5.1",
    "pytest>=9.0.2",
    "python-dotenv>=1.2.2",
    "respx>=0.22.0",
    "ruff>=0.15.4",
]
